requirements.txt
Django>=4.2,<5.0
psycopg2-binary
redis
django-redis
celery
bamboo-engine>=3.0.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Werkzeug>=1.0.0
pyparsing>=2.2.0
Mako>=1.1.4
prometheus-client>=0.9.0
networkx>=2.4
django-mptt>=0.11.0
requests>=2.22.0
django-celery-beat>=2.1.0
jsonschema>=2.5.1
ujson>=4.0.0
django-timezone-field>=4.0
boto3>=1.9.130
isodate>=0.6.0
django-mptt-admin
django-filter
google-genai>=1.0.0
Pillow>=10.0.0
openai>=1.0.0
beautifulsoup4>=4.10.0
python-telegram-bot>=20.0
channels>=4.0.0
channels-redis>=4.0.0
daphne>=4.0.0